recipes-app/measy-iot/files/usr/share/measy_iot/requirements.txt
Babel==2.6.0
backports-abc==0.5
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==0.12.2
Flask-AppBuilder==1.12.3
Flask-Babel==0.11.1
Flask-Cors==3.0.3
Flask-JSONRPC==0.3.1
Flask-Login==0.4.0
Flask-OpenID==1.2.5
Flask-SocketIO==3.3.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==1.2.2
greenlet==0.4.12
ifparser==0.8.2
itsdangerous==0.24
iw-parse==0.0.2
Jinja2==2.10
MarkupSafe==1.0
paho-mqtt==1.4.0
python-dateutil==2.7.5
python-engineio==3.5.1
python-openid==2.2.5
python-socketio==3.1.2
python-tdbus==0.11
pytz==2018.7
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.11.0
smbus==1.1
SQLAlchemy==1.2.2
uWSGI==2.0.15
Werkzeug==0.12.2
WTForms==2.2.1